requirements.txt
beautifulsoup4==4.8.1
Django==2.2.6
django-bootstrap4==1.0.1
django-ckeditor==5.7.1
django-js-asset==1.2.2
Pillow==6.2.0
pytz==2019.3
soupsieve==1.9.4
sqlparse==0.3.0
